requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.99
PyAudio==0.2.14
python-dotenv==1.0.1
pytest==8.3.5
pytest-cov==5.0.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
Flask==3.0.3
ruff==0.6.9
